Use 32-bit EMAC counters and add counters module
Change Counters fields from uint64_t to uint32_t and adapt callers. Introduces emac_counters.h and emac_counters.cpp to expose emac::eth::globals::counter and map hardware counters into network::iface::GetCounters(). Initialize the counter in Start(), increment sent/received/received_error in emac_eth.cpp, and update JSON status code to write 32-bit counters (U32) and string helpers. Also adjust the Linux parser to read 32-bit values and add file headers/license to the new source file.

+++ lib-network/include/network_iface.h
@@ -92,10 +92,10 @@ inline char AddressingMode() {
 
 struct Counters {
     struct Receive {
-        uint64_t ok = 0, err = 0, drp = 0, ovr = 0;
+        uint32_t ok = 0, err = 0, drp = 0, ovr = 0;
     } rx;
     struct Transmit {
-        uint64_t ok = 0, err = 0, drp = 0, ovr = 0;
+        uint32_t ok = 0, err = 0, drp = 0, ovr = 0;
     } tx;
 };
